requirements.txt
requests
beautifulsoup4
yt-dlp
selenium
webdriver_manager
tqdm
python-telegram-bot
pyrogram
tgcrypto
python-dotenv
pillow
arabic-reshaper
python-bidi
